# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SPM API
  version: 1.0.0
  description: Senior Project Management system API
servers:
  - url: /api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorMessage:
      type: object
      properties:
        message:
          type: string
...
    GroupMember:
      type: object
      properties:
        studentId:
          type: string
        role:
          type: string
          enum: [TEAM_LEADER, MEMBER]
        joinedAt:
          type: string
          format: date-time
...
    GroupDetailResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        groupName:
          type: string
        termId:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        jiraSpaceUrl:
          type: string
          nullable: true
        jiraProjectKey:
          type: string
          nullable: true
        jiraBound:
          type: boolean
        githubOrgName:
          type: string
          nullable: true
        githubBound:
          type: boolean
        advisorId:
          type: string
          format: uuid
          nullable: true
        members:
          type: array
          items:
            $ref: '#/components/schemas/GroupMember'
...
    AdvisorRequestPayload:
      type: object
      required: [advisorId]
      properties:
        advisorId:
          type: string
          format: uuid
    AdvisorRequestCreatedResponse:
      type: object
      properties:
        requestId:
          type: string
          format: uuid
        groupId:
          type: string
          format: uuid
        advisorId:
          type: string
          format: uuid
        status:
          type: string
        sentAt:
          type: string
          format: date-time
    AdvisorRequestDetailStudent:
      type: object
      properties:
        requestId:
          type: string
          format: uuid
        advisorId:
          type: string
          format: uuid
        advisorName:
          type: string
        status:
          type: string
        sentAt:
          type: string
          format: date-time
        respondedAt:
          type: string
          format: date-time
          nullable: true
    AdvisorRequestCancelResponse:
      type: object
      properties:
        requestId:
          type: string
          format: uuid
        status:
          type: string
...
    AdvisorRequestDetailProfessor:
      type: object
      properties:
        requestId:
          type: string
          format: uuid
        group:
          $ref: '#/components/schemas/GroupDetailResponse'
        sentAt:
          type: string
          format: date-time
    RespondAdvisorRequestPayload:
      type: object
      required: [accept]
      properties:
        accept:
          type: boolean
    RespondAdvisorRequestAcceptResponse:
      type: object
      properties:
        requestId:
          type: string
          format: uuid
        status:
          type: string
        groupId:
          type: string
          format: uuid
        groupStatus:
          type: string
    RespondAdvisorRequestRejectResponse:
      type: object
      properties:
        requestId:
          type: string
          format: uuid
        status:
          type: string
...
security:
  - bearerAuth: []
paths:
...
  /groups/{groupId}/advisor-request:
    post:
      tags: [Student]
      summary: Send an advisor request
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdvisorRequestPayload'
      responses:
        '201':
          description: Advisor request sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdvisorRequestCreatedResponse'
        '400':
          description: Window inactive, status not TOOLS_BOUND, or advisor at capacity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '403':
          description: Caller is not the Team Leader
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '404':
          description: Advisor not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '409':
          description: Pending request already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
    get:
      tags: [Student]
      summary: Get group's advisor request status
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Returns the latest advisor request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdvisorRequestDetailStudent'
        '403':
          description: Not a member of the group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '404':
          description: No request found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
    delete:
      tags: [Student]
      summary: Cancel an advisor request
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Request cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdvisorRequestCancelResponse'
        '400':
          description: No active pending request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '403':
          description: Caller is not the Team Leader
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '404':
          description: Group or request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
...
  /advisor/requests/{requestId}:
    get:
      tags: [Professor]
      summary: Get details of a specific advisor request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Returns request details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdvisorRequestDetailProfessor'
        '403':
          description: Request not addressed to this professor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '404':
          description: Request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
  /advisor/requests/{requestId}/respond:
    patch:
      tags: [Professor]
      summary: Respond to an advisor request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RespondAdvisorRequestPayload'
      responses:
        '200':
          description: Request processed successfully
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/RespondAdvisorRequestAcceptResponse'
                  - $ref: '#/components/schemas/RespondAdvisorRequestRejectResponse'
        '400':
          description: Request is no longer pending; or advisor has reached maximum capacity (AdvisorAtCapacityException — request stays PENDING, group can retry)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '403':
          description: Request not addressed to this professor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '404':
          description: Request not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
...